# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jack Pine Farm Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin, orders]
      summary: List all orders (admin)
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/OrderStatus"
          description: Filter by status
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}:
    get:
      operationId: adminGetOrder
      tags: [admin, orders]
      summary: Get order details (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin, orders]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/refund-giblets:
    post:
      operationId: adminRefundGiblets
      tags: [admin, orders]
      summary: Refund the giblets add-on for an order (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Cannot refund (no giblets or already refunded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/notes:
    post:
      operationId: adminAddOrderNote
      tags: [admin, orders]
      summary: Add a note to an order (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddNoteBody"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/events:
    get:
      operationId: adminGetOrderEvents
      tags: [admin, orders]
      summary: Get audit trail for an order (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderEvent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/orders/{id}/assign-batch:
    patch:
      operationId: adminAssignOrderBatch
      tags: [admin, orders]
      summary: Assign an order to a preorder batch (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignBatchBody"
      responses:
        "200":
          description: Batch assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/customers:
    get:
      operationId: adminListCustomers
      tags: [admin, customers]
      summary: List all customers (admin)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCustomerSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/customers/{id}:
    get:
      operationId: adminGetCustomer
      tags: [admin, customers]
      summary: Get customer detail with order history (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PricingType:
      type: string
      enum:
        - unit
        - deposit
    OrderStatus:
      type: string
      enum:
        - pending_payment
        - deposit_paid
        - cash_pending
        - pickup_assigned
        - weights_entered
        - invoice_sent
        - fulfilled
        - cancelled
        - no_show
    PaymentMethod:
      type: string
      enum:
        - stripe
        - cash
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        quantity:
          type: integer
        pricingType:
          $ref: "#/components/schemas/PricingType"
        unitPriceInCents:
          type: integer
        unitLabel:
          type: ["string", "null"]
        isGiblets:
          type: boolean
        lineTotalInCents:
          type: integer
      required:
        - id
        - productId
        - productName
        - quantity
        - pricingType
        - unitPriceInCents
        - unitLabel
        - isGiblets
        - lineTotalInCents
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        totalInCents:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - status
        - paymentMethod
        - totalInCents
        - createdAt
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        stripeCheckoutSessionId:
          type: ["string", "null"]
        totalInCents:
          type: integer
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - customerName
        - customerEmail
        - customerPhone
        - status
        - paymentMethod
        - stripeCheckoutSessionId
        - totalInCents
        - notes
        - items
        - createdAt
        - updatedAt
...
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        note:
          type: string
      required:
        - status
    AddNoteBody:
      type: object
      properties:
        body:
          type: string
      required:
        - body
    AssignBatchBody:
      type: object
      properties:
        batchId:
          type: ["integer", "null"]
      required:
        - batchId
    OrderEventType:
      type: string
      enum:
        - note
        - status_change
        - refund
        - invoice_sent
        - pickup_assigned
        - weights_entered
    OrderEvent:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        eventType:
          $ref: "#/components/schemas/OrderEventType"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - eventType
        - body
        - createdAt
    AdminCustomerSummary:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - phone
        - orderCount
        - createdAt
    AdminCustomerDetail:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        orderCount:
          type: integer
        orders:
          type: array
          items:
            $ref: "#/components/schemas/OrderSummary"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - phone
        - emailVerified
        - orderCount
        - orders
        - createdAt
...