# lib/api-zod/src/generated/types/adminCustomerDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSummary } from "./orderSummary";

export interface AdminCustomerDetail {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone: string | null;
  emailVerified: boolean;
  orderCount: number;
  orders: OrderSummary[];
  createdAt: Date;
}