# lib/api-zod/src/generated/types/adminCustomerSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCustomerSummary {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone: string | null;
  orderCount: number;
  createdAt: Date;
}