# lib/api-zod/src/generated/types/orderEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderEventType } from "./orderEventType";

export interface OrderEvent {
  id: number;
  orderId: number;
  eventType: OrderEventType;
  body: string;
  createdAt: Date;
}