# lib/api-zod/src/generated/types/orderEventType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export type OrderEventType =
  (typeof OrderEventType)[keyof typeof OrderEventType];

export const OrderEventType = {
  note: "note",
  status_change: "status_change",
  refund: "refund",
  invoice_sent: "invoice_sent",
  pickup_assigned: "pickup_assigned",
  weights_entered: "weights_entered",
} as const;