# lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jack Pine Farm Store API
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending_payment: "pending_payment",
  deposit_paid: "deposit_paid",
  cash_pending: "cash_pending",
  pickup_assigned: "pickup_assigned",
  weights_entered: "weights_entered",
  invoice_sent: "invoice_sent",
  fulfilled: "fulfilled",
  cancelled: "cancelled",
  no_show: "no_show",
} as const;